force data to double precision before computing matrix rank
git-svn-id: https://sccn.ucsd.edu/svn/software/eeglab@9632 1e58238c-3a72-4c0f-aabe-36d030380f39

+++ functions/popfunc/pop_runica.m
@@ -353,7 +353,7 @@ end;
 % compute ICA on a definite set
 % -----------------------------
 tmpdata = reshape( EEG.data(g.chanind,:,:), length(g.chanind), EEG.pnts*EEG.trials);
-tmprank = getrank(tmpdata(:,1:min(3000, size(tmpdata,2))));
+tmprank = getrank(double(tmpdata(:,1:min(3000, size(tmpdata,2)))));
 tmpdata = tmpdata - repmat(mean(tmpdata,2), [1 size(tmpdata,2)]); % zero mean 
 if ~strcmpi(lower(g.icatype), 'binica')
     try
